call cast function on ids to pass the test

+++ lib/src/resouces/news_api_provider.dart
@@ -13,7 +13,8 @@ class NewsApiProvider {
     final response = await client.get('$_root/topstories.json');
     final ids = json.decode(response.body);
 
-    return ids;
+    // indicate dart that ids is a list of integer
+    return ids.cast<int>();
   }
 
   Future<ItemModel> fetchItem(int id) async {
